Hugging-face/requirements.txt
gradio
modal
transformers
torch
elevenlabs
mcp
openai-whisper
pillow
scipy
numpy
accelerate
sentencepiece
protobuf
